// LDC.Atlas.WebApp/Client/app/shared/entities/company-setup.entity.ts
export class CompanySetup {
    id: number;
    companyName: string;
    companyId: string;
    companyFriendlyCode: string;
    companyTypeId: number;
    legalEntityCode: string;
    legalEntity: string;
    functionalCurrencyCode: string;
    statutoryCurrencyCode: string;
    companyPlatformId: number;
    countryId: number;
    timeZone: string;
    ldcRegionId: number;
    weightUnitId: number;
    counterpartyId: number;
    companyDate: Date;
    cropYearFormatId: number;
    defaultBrokerId: number;
    defaultProvinceId: number;
    defaultBranchId: number;
    isProvinceEnable: boolean;
    priceUnitId: number;
}